+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Remove unnecessary code in overdue screen
 - Update call result outcome design in contact patient bottom sheet
 - Add support for select and download in overdue search screen
+- Add support for select and share in overdue search screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Changes